refactor: 責務を分けるため、引数をクラス化し、そちらを使用するようにした
